Fix #195 -- Playerconnection subclasses erroring
If another plugin overwrote the PlayerConnection class, getting the
networkManager requires also looking through the superclasses. This
patch makes the getValue method (which was used for this) also look
through all superclasses, in order to find the field.

Whether this change should be applied more generally would need some
more thought as to where it is used.