Fix #229, a NBT hierarchy change
== Problem ==

Starting with some newer 1.13 builds, the NBT hierarchy has changed.
`NBTBase` used to be a *superclass* of `NBTTagCompound` and the
comphenix class used this relationship to find the base class.
Now `NBTTagCompound` *implements* the *interface* `NBTBase`, which
makes the `getSuperclass` check performed return `Object`.
This causes the subsequent check for the `getTypeId` method to fail.

== Mitigation ==

The `NBTBase` class is now looked up via the Reflector mechanism already
present in OCM. This mechanism directly resolves the runtime path of the
NMS class and returns it, thus circumventing the checks.