Prevent sweep particle packets from being sent
Sweeps already did no damage, but now the visual indication of one
should be gone as well.

It works by simply cancelling the packet if it is:
a) A particle packet
b) Conveys information for the "Sweep" packet

This is done via two different adapters in order to work on 1.13 and up
and also on 1.9 to 1.12.2.