Fix a MC version conflict in ModuleFishingKnockback
Version 1.13 defines "FishingEvent#getHook" to return a "FishHook",
in lower versions it returns a "Fish" instead. This caused the compiled
code to fail on lower version, as it targeted the 1.13 method.

This commit introduces a feature branch that either just uses the 1.13
method or falls back to reflection (caching the Method to somewhat
lessen the performance impact) on lower versions.

Ideally you would have a part that is compiled against 1.12.2, but that
is not nice to build, as the namespace of the two APIs is exactly the
same.