Fix continuous golden apple eating
The Client and server go out of sync causing the client to keep eating
forever as it does not realise the server cancelled the action. This is
probably a bukkit/spigot bug?
To circumvent that we force-update the player's inventory.

Doing that feels a bit heavy-handed, but it works, is not triggered too
frequently and does not require digging through packet captures for an
hour, which is a plus.

Closes: #418