Provide some logic for 1.17 packets, disregarding immutability
This will need to be addressed by rewriting the API to take a packet and
produce a new one. This change can be easily made to the old APIs as
well, so at least that will be consistent.

However, all packet modification will now need to clone the packet using
adequate constructors.