Remove redundant overdamage
Minecraft already has existing overdamage logic that compares incoming
damage to the last damage and adjusts the final damage when the incoming
damage is larger than the last damage during invulnerability ticks.

The OCM overdamage logic does not overwrite the existing overdamage logic
and triggers strange interactions when entity.setLastDamage() is called.

OCM should rely on the existing overdamage logic instead of layering on top of it.

The vanilla overdamage logic can be found in the LivingEntity class's override
of the Entity#hurt(DamageSource,float) method.

https://mappings.dev/1.21.3/net/minecraft/world/entity/Entity.html